Optimizing by changing from getComputers to getNodes.
In environments with hundreds of slaves, hundred of builds in the queue
to be throttled and hundreds of simultaneous web requests.
Jenkins then spends a lot of time in the unnecessary effort to slowly
sort the list of computers for this plugin to count the number of builds
causing the web requests to wait for the queue to release its monitor
so they can render it on the page giving the impression that the master
is overloaded while in fact it is just waiting.
Changing to getNodes() should speed responsiveness up a bit.

Change-Id: I09e0f2b2f2d94fc4366ed84b2e51ecaea961efc1