This change fixes an issue introduced with the Matrix support around allowing flyweight tasks to throttle a given category.
Essentially, if a non-Matrix flyweight task in category A was running, we wouldn't count its one-off executor when determining whether another project in category A should be scheduled.

The fix is to count all executors for all tasks that should be throttled instead of only counting one-off executors for Matrix jobs.