[JENKINS-28133] Prevent concurrent builds from exceeding max of category
When all maximum concurrent builds for a category are being consumed,
multiple queued builds from that category can start when only one
finishes.  This causes the maximum concurrent builds for the category to
be exceeded, so long as executors on the node were also available.

Implements an additional cause to block execution if the task is not at
the top of the category's queue, allowing only one task to be scheduled
at a time.