Add tests for existing functionality
This plugin contains tests for Freestyle jobs in ThrottleIntegrationTest
and tests for Pipeline jobs in ThrottleStepTest.

ThrottleIntegrationTest contains a test for jobs that throttle the
number of concurrent builds per node for a category; however, there is
no test for jobs that throttle the total number of concurrent builds for
a category. This commit adds such a test.

ThrottleStepTest contains a test for node-level throttling with a
`throttle` step; however, there is no test for job-level throttling.
This commit adds such a test as well.