Use FlowExecutionList to calculate the number of running pipeline jobs
When throttling the number of builds of a given pipeline job, the
current code iterates through all flyweight executors on the system.
This interacts poorly with other plugins, which leak flyweight executors
at the present time. It's also not efficient. Instead, we use
FlowExecutionList to calculate the number of builds of a given pipeline
job. The old code path remains in place for non-pipeline jobs.

Incidentally, before this change running pipeline jobs would not be
throttled correctly after a Jenkins restart. This functionality now
works, so this commit adds a test to ensure we don't regress.