Use FlowExecutionList to calculate the number of running Pipeline jobs (#57)
Use FlowExecutionList to calculate the number of running Pipeline jobs

When throttling the number of builds of a given Pipeline job, the
current code iterates through all flyweight executors on the system.
This interacts poorly with other plugins, which leak flyweight executors
at the present time. It's also not efficient. Instead, we use
FlowExecutionList to calculate the number of builds of a given Pipeline
job. The old code path remains in place for non-Pipeline jobs.