Add PauseAction when throttling pipelines
The PauseAction is used by pipeline visualisers (such as
pipeline-stage-view) to indicate when a stage has been paused, and
for how long is was staged for. This can improve the display of
stages timings, as the paused time is removed from the overall stage
time.

The PauseAction only applies to stages, so only applies when the
throttle step is used in conjunction with a pipeline stage.